Add base items, Item highlighting
There is a class for items now, just drop the BP_Base Interactable into
the world and then set its interactableMesh to the weapon that you want.
Hovering over the item will highlight it, indicating that you can pick
it up. As of right now you can't pick it up or interact with it at all.