1, 2 PUNCH! Add punch animations, start of damage calc
Our character can now throw some mean punches. The punches also do a
trace for targets to hit but doesn't do anything with that information.
Because of the way i've implemented punching the code looks a little
spaghettish... But the damage calc also works for all weapons too. They
won't be usable for the weapons until they get animations though.